#36 change UI based on Feedback
