+++ lib/view/web_display/about_pages_web/about_two_web.dart
@@ -31,33 +31,33 @@ class AboutTwo extends StatelessWidget {
               HeightSizedBox(targetSize: deviceHeight, value: 0.03),
               BodyText(
                 text: "「何かわからないけど、このデザインいいな」「使いやすくて楽しい、ワクワクする」",
-                color: Colors.black,
+                color: const Color.fromRGBO(0, 0, 0, 0.8),
                 fontFamily: 'Noto Sans JP',
-                fontSize: deviceHeight * 0.023,
+                fontSize: deviceHeight * 0.028,
                 fontWeight: FontWeight.normal,
               ),
               HeightSizedBox(targetSize: deviceHeight, value: 0.005),
               BodyText(
                 text: "良いデザインとは、自然と人の興味を惹き、魅了する力があると私は思っています。",
-                color: Colors.black,
+                color: const Color.fromRGBO(0, 0, 0, 0.8),
                 fontFamily: 'Noto Sans JP',
-                fontSize: deviceHeight * 0.023,
+                fontSize: deviceHeight * 0.028,
                 fontWeight: FontWeight.normal,
               ),
               HeightSizedBox(targetSize: deviceHeight, value: 0.005),
               BodyText(
                 text: "サービス・プロダクトを使ってくれる人のことを最優先に考え、デザインすることで",
-                color: Colors.black,
+                color: const Color.fromRGBO(0, 0, 0, 0.8),
                 fontFamily: 'Noto Sans JP',
-                fontSize: deviceHeight * 0.023,
+                fontSize: deviceHeight * 0.028,
                 fontWeight: FontWeight.normal,
               ),
               HeightSizedBox(targetSize: deviceHeight, value: 0.005),
               BodyText(
                 text: "誰もがデザインを考える世界を、より多くの人にデザインの力で広げていくことが私のミッションです。",
-                color: Colors.black,
+                color: const Color.fromRGBO(0, 0, 0, 0.8),
                 fontFamily: 'Noto Sans JP',
-                fontSize: deviceHeight * 0.023,
+                fontSize: deviceHeight * 0.028,
                 fontWeight: FontWeight.normal,
               ),
             ],
