+++ lib/view/web_display/works_pages_web/shusseki_pages_web/shusseki_five_web.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../../../../../utils/extract_widget.dart';
 import '../../../../utils/project_widget/works_widget.dart';
 
-// UI改修背景
+// 開発での感想
 class ShussekiFive extends StatelessWidget {
   const ShussekiFive({
     Key? key,
@@ -21,12 +21,137 @@ class ShussekiFive extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: <Widget>[
-          Column(
+          Row(
             mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.center,
             children: [
+              Padding(
+                padding: EdgeInsets.symmetric(
+                  horizontal: deviceHeight * 0.03,
+                ),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    SizedBox(
+                      height: deviceHeight * 0.3,
+                      child: Container(
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(10),
+                          color: const Color.fromRGBO(151, 151, 151, 0.1),
+                        ),
+                        child: Padding(
+                          padding: EdgeInsets.all(deviceHeight * 0.03),
+                          child: Column(
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            crossAxisAlignment: CrossAxisAlignment.center,
+                            children: [
+                              BodyText(
+                                text: "なぜFlutterなのか？",
+                                color: Colors.black,
+                                fontSize: deviceHeight * 0.025,
+                                fontWeight: FontWeight.bold,
+                                fontFamily: "Noto Sans JP",
+                              ),
+                              HeightSizedBox(
+                                  targetSize: deviceHeight, value: 0.025),
+                              BodyText(
+                                text: "元々Swiftをやっていて、アプリ開発に興味があった",
+                                color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                fontSize: deviceHeight * 0.02,
+                                fontWeight: FontWeight.normal,
+                                fontFamily: "Noto Sans JP",
+                              ),
+                              HeightSizedBox(
+                                  targetSize: deviceHeight, value: 0.02),
+                              HighPaddingText(
+                                text:
+                                    "クロスプラットフォームで\niOS・Android・Webなど多様な開発を可能とし\nWidgetが豊富に揃っており、学習コストが低いため",
+                                color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                fontSize: deviceHeight * 0.02,
+                                fontWeight: FontWeight.normal,
+                                fontFamily: "Noto Sans JP",
+                                textAlign: TextAlign.center,
+                                paddingValue: 1.3,
+                              ),
+                              HeightSizedBox(
+                                  targetSize: deviceHeight, value: 0.02),
+                              BodyText(
+                                text: "人気の言語というのもあり、成長度合いを感じたため",
+                                color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                fontSize: deviceHeight * 0.02,
+                                fontWeight: FontWeight.normal,
+                                fontFamily: "Noto Sans JP",
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                    HeightSizedBox(targetSize: deviceHeight, value: 0.03),
+                    SizedBox(
+                      height: deviceHeight * 0.3,
+                      child: Container(
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(10),
+                          color: const Color.fromRGBO(151, 151, 151, 0.1),
+                        ),
+                        child: Padding(
+                          padding: EdgeInsets.all(deviceHeight * 0.03),
+                          child: Column(
+                            mainAxisAlignment: MainAxisAlignment.center,
+                            crossAxisAlignment: CrossAxisAlignment.center,
+                            children: [
+                              BodyText(
+                                text: "開発で難しかった点",
+                                color: Colors.black,
+                                fontSize: deviceHeight * 0.025,
+                                fontWeight: FontWeight.bold,
+                                fontFamily: "Noto Sans JP",
+                              ),
+                              HeightSizedBox(
+                                  targetSize: deviceHeight, value: 0.025),
+                              BodyText(
+                                text: "データベースのデータを非同期処理で保存・受け取る",
+                                color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                fontSize: deviceHeight * 0.02,
+                                fontWeight: FontWeight.normal,
+                                fontFamily: "Noto Sans JP",
+                              ),
+                              HeightSizedBox(
+                                  targetSize: deviceHeight, value: 0.02),
+                              HighPaddingText(
+                                text: "リリースまわりの設定や\nFlutter以外でのエラーに苦戦しました",
+                                color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                fontSize: deviceHeight * 0.02,
+                                fontWeight: FontWeight.normal,
+                                fontFamily: "Noto Sans JP",
+                                textAlign: TextAlign.center,
+                                paddingValue: 1.3,
+                              ),
+                              HeightSizedBox(
+                                  targetSize: deviceHeight, value: 0.02),
+                              HighPaddingText(
+                                text:
+                                    "情報が少ないので海外の記事を読んだり\n質問サイトで英語での質問をする必要があった",
+                                color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                fontSize: deviceHeight * 0.02,
+                                fontWeight: FontWeight.normal,
+                                fontFamily: "Noto Sans JP",
+                                textAlign: TextAlign.center,
+                                paddingValue: 1.3,
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              WidthSizedBox(targetSize: deviceWidth, value: 0.015),
               SizedBox(
-                width: deviceWidth * 0.8,
+                width: deviceWidth * 0.4,
+                height: deviceHeight * 0.6,
                 child: Container(
                   decoration: BoxDecoration(
                     borderRadius: BorderRadius.circular(10),
@@ -41,224 +166,163 @@ class ShussekiFive extends StatelessWidget {
                     color: Colors.white,
                   ),
                   child: Padding(
-                    padding: EdgeInsets.all(deviceHeight * 0.015),
-                    child: Row(
+                    padding: EdgeInsets.only(
+                        top: deviceHeight * 0.03,
+                        bottom: deviceHeight * 0.02,
+                        left: deviceHeight * 0.03,
+                        right: deviceHeight * 0.03),
+                    child: Column(
                       mainAxisAlignment: MainAxisAlignment.center,
-                      crossAxisAlignment: CrossAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
                         Column(
                           crossAxisAlignment: CrossAxisAlignment.start,
                           children: [
                             BodyText(
-                              text: "UI改修背景",
+                              text: "現状(2022年5月-12月)",
                               color: const Color(0xFF379BA5),
                               fontSize: deviceHeight * 0.028,
                               fontWeight: FontWeight.bold,
                               fontFamily: "源ノ角ゴシック VF",
                             ),
                             Padding(
-                              padding: EdgeInsets.all(deviceHeight * 0.03),
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
+                              padding: EdgeInsets.all(deviceHeight * 0.02),
+                              child: Row(
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
                                 children: [
-                                  BodyText(
-                                    text: "「OOUIってなんだ!?」",
-                                    color: Colors.black,
-                                    fontSize: deviceHeight * 0.03,
-                                    fontWeight: FontWeight.bold,
-                                    fontFamily: "源ノ角ゴシック VF",
+                                  Column(
+                                    children: [
+                                      BodyText(
+                                        text: "ユーザー総数",
+                                        color: Colors.black,
+                                        fontSize: deviceHeight * 0.02,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                      HeightSizedBox(
+                                          targetSize: deviceHeight,
+                                          value: 0.01),
+                                      BodyText(
+                                        text: "126",
+                                        color: Colors.black,
+                                        fontSize: deviceHeight * 0.04,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                    ],
                                   ),
-                                  HeightSizedBox(
-                                      targetSize: deviceHeight, value: 0.01),
-                                  BodyText(
-                                    text:
-                                        "2022年頃からUI/UXに興味を抱き始め、授業でもUI/UXコースに所属しました。",
-                                    color: const Color.fromRGBO(0, 0, 0, 0.6),
-                                    fontSize: deviceHeight * 0.02,
-                                    fontWeight: FontWeight.normal,
-                                    fontFamily: "Noto Sans JP",
+                                  Column(
+                                    children: [
+                                      BodyText(
+                                        text: "入館総数(43団体)",
+                                        color: Colors.black,
+                                        fontSize: deviceHeight * 0.02,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                      HeightSizedBox(
+                                          targetSize: deviceHeight,
+                                          value: 0.01),
+                                      BodyText(
+                                        text: "459",
+                                        color: Colors.black,
+                                        fontSize: deviceHeight * 0.04,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                    ],
                                   ),
-                                  BodyText(
-                                    text:
-                                        "授業や自主学習を行っていくうちに、様々な新しい考え方に触れ、ただ作るだけではなく",
-                                    color: const Color.fromRGBO(0, 0, 0, 0.6),
-                                    fontSize: deviceHeight * 0.02,
-                                    fontWeight: FontWeight.normal,
-                                    fontFamily: "Noto Sans JP",
-                                  ),
-                                  BodyText(
-                                    text:
-                                        "5段階モデルを意識した開発を行うことで、シュッ席のデザインの悪さにも気づきました。",
-                                    color: const Color.fromRGBO(0, 0, 0, 0.6),
-                                    fontSize: deviceHeight * 0.02,
-                                    fontWeight: FontWeight.normal,
-                                    fontFamily: "Noto Sans JP",
+                                  Column(
+                                    children: [
+                                      BodyText(
+                                        text: "アクティブユーザー(月)",
+                                        color: Colors.black,
+                                        fontSize: deviceHeight * 0.02,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                      HeightSizedBox(
+                                          targetSize: deviceHeight,
+                                          value: 0.01),
+                                      BodyText(
+                                        text: "25",
+                                        color: Colors.black,
+                                        fontSize: deviceHeight * 0.04,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                    ],
                                   ),
                                 ],
                               ),
                             ),
+                          ],
+                        ),
+                        HeightSizedBox(targetSize: deviceHeight, value: 0.025),
+                        Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            BodyText(
+                              text: "課題点",
+                              color: const Color(0xFF379BA5),
+                              fontSize: deviceHeight * 0.028,
+                              fontWeight: FontWeight.bold,
+                              fontFamily: "源ノ角ゴシック VF",
+                            ),
                             Padding(
-                              padding:
-                                  EdgeInsets.only(left: deviceWidth * 0.02),
+                              padding: EdgeInsets.all(deviceHeight * 0.02),
                               child: Column(
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
-                                  ShadowContainerText(
-                                    deviceHeight: deviceHeight,
-                                    text: '使う人のことを考えれずに、開発を進めてしまった',
+                                  BodyText(
+                                    text: "ユーザー数は増えても、アクティブユーザーがいない",
+                                    color: Colors.black,
+                                    fontSize: deviceHeight * 0.025,
+                                    fontWeight: FontWeight.bold,
+                                    fontFamily: "Noto Sans JP",
                                   ),
                                   HeightSizedBox(
-                                      targetSize: deviceHeight, value: 0.01),
-                                  ShadowContainerText(
-                                    deviceHeight: deviceHeight,
-                                    text: 'オブジェクトではなく、タスクベースでの開発だった',
+                                      targetSize: deviceHeight, value: 0.013),
+                                  HighPaddingText(
+                                    text:
+                                        "入館するまでのフローが複雑なため、ユーザーが機能を使うまでにアプリから離脱してしまっている。",
+                                    color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                    fontSize: deviceHeight * 0.02,
+                                    fontWeight: FontWeight.normal,
+                                    fontFamily: "Noto Sans JP",
+                                    textAlign: TextAlign.start,
+                                    paddingValue: 1.3,
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.03),
+                                  BodyText(
+                                    text: "機能の使い方が一目で分からない",
+                                    color: Colors.black,
+                                    fontSize: deviceHeight * 0.025,
+                                    fontWeight: FontWeight.bold,
+                                    fontFamily: "Noto Sans JP",
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.013),
+                                  HighPaddingText(
+                                    text:
+                                        "押してみるまでどうなるのかが予想できないUIなために、ユーザー任せなサービスとなってしまっている。\n例外処理も欠点が多く、ユーザーが離れる原因となっている。",
+                                    color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                    fontSize: deviceHeight * 0.02,
+                                    fontWeight: FontWeight.normal,
+                                    fontFamily: "Noto Sans JP",
+                                    textAlign: TextAlign.start,
+                                    paddingValue: 1.3,
                                   ),
                                 ],
                               ),
                             )
                           ],
                         ),
-                        const ImageWidget(
-                          heightValue: 0.4,
-                          imagePath: "https://i.imgur.com/htHfPUH.png",
-                        ),
                       ],
                     ),
                   ),
-                ),
-              ),
-              HeightSizedBox(targetSize: deviceHeight, value: 0.02),
-              Padding(
-                padding: EdgeInsets.only(
-                    top: deviceHeight * 0.03,
-                    right: deviceHeight * 0.03,
-                    left: deviceHeight * 0.03),
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    WidthSizedBox(targetSize: deviceWidth, value: 0.03),
-                    SizedBox(
-                      height: deviceHeight * 0.3,
-                      child: Container(
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color.fromRGBO(151, 151, 151, 0.1),
-                        ),
-                        child: Padding(
-                          padding: EdgeInsets.all(deviceHeight * 0.03),
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
-                              BodyText(
-                                text: "「シュッ席」開発時の自分",
-                                color: Colors.black,
-                                fontSize: deviceHeight * 0.03,
-                                fontWeight: FontWeight.bold,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                              HeightSizedBox(
-                                  targetSize: deviceHeight, value: 0.02),
-                              BodyText(
-                                text: "・何の機能が必要だろうか？",
-                                color: const Color(0xFF379BA5),
-                                fontSize: deviceHeight * 0.02,
-                                fontWeight: FontWeight.normal,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                              HeightSizedBox(
-                                  targetSize: deviceHeight, value: 0.02),
-                              BodyText(
-                                text: "・前提さえ理解すれば使ってくれるはず",
-                                color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                fontSize: deviceHeight * 0.02,
-                                fontWeight: FontWeight.normal,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                              HeightSizedBox(
-                                  targetSize: deviceHeight, value: 0.02),
-                              BodyText(
-                                text: "・ターゲットさえ決めれば問題ないだろう",
-                                color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                fontSize: deviceHeight * 0.02,
-                                fontWeight: FontWeight.normal,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                            ],
-                          ),
-                        ),
-                      ),
-                    ),
-                    WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                    Column(
-                      children: [
-                        const ImageWidget(
-                          heightValue: 0.2,
-                          imagePath: "https://i.imgur.com/7BDf19N.png",
-                        ),
-                        BodyText(
-                          text: "→",
-                          color: const Color.fromRGBO(151, 151, 151, 1.0),
-                          fontSize: deviceHeight * 0.1,
-                          fontWeight: FontWeight.normal,
-                          fontFamily: "Noto Sans JP",
-                        ),
-                      ],
-                    ),
-                    WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                    SizedBox(
-                      height: deviceHeight * 0.3,
-                      child: Container(
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(10),
-                          color: const Color.fromRGBO(151, 151, 151, 0.1),
-                        ),
-                        child: Padding(
-                          padding: EdgeInsets.all(deviceHeight * 0.03),
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: [
-                              BodyText(
-                                text: "UI/UXを学び始めた自分",
-                                color: Colors.black,
-                                fontSize: deviceHeight * 0.03,
-                                fontWeight: FontWeight.bold,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                              HeightSizedBox(
-                                  targetSize: deviceHeight, value: 0.02),
-                              BodyText(
-                                text: "・要件に含まれるオブジェクトは何だろうか？",
-                                color: const Color(0xFF379BA5),
-                                fontSize: deviceHeight * 0.02,
-                                fontWeight: FontWeight.normal,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                              HeightSizedBox(
-                                  targetSize: deviceHeight, value: 0.02),
-                              BodyText(
-                                text: "・ユーザーが真に求めているものは何だろう？",
-                                color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                fontSize: deviceHeight * 0.02,
-                                fontWeight: FontWeight.normal,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                              HeightSizedBox(
-                                  targetSize: deviceHeight, value: 0.02),
-                              BodyText(
-                                text: "・ビジネス戦略より、まずは人間中心の開発に",
-                                color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                fontSize: deviceHeight * 0.02,
-                                fontWeight: FontWeight.normal,
-                                fontFamily: "Noto Sans JP",
-                              ),
-                            ],
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
                 ),
               ),
             ],
