+++ lib/view/web_display/works_pages_web/shusseki_pages_web/shusseki_seven_web.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../../../../utils/extract_widget.dart';
 
-// UI改修後　UI
+// UI設計
 class ShussekiSeven extends StatelessWidget {
   const ShussekiSeven({
     Key? key,
@@ -25,7 +25,7 @@ class ShussekiSeven extends StatelessWidget {
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               BodyText(
-                text: "UIの変化",
+                text: "UI設計",
                 color: const Color(0xFF379BA5),
                 fontSize: deviceHeight * 0.035,
                 fontWeight: FontWeight.bold,
@@ -40,127 +40,47 @@ class ShussekiSeven extends StatelessWidget {
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
                         BodyText(
-                          text: "Before",
+                          text: "ユーザーフローの確認",
                           color: Colors.black,
                           fontSize: deviceHeight * 0.03,
                           fontWeight: FontWeight.bold,
                           fontFamily: "Noto Snas JP",
                         ),
-                        Row(
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/sPpXecU.png",
-                            ),
-                            WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/BYjCQss.png",
-                            ),
-                            WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/ihV37xj.png",
-                            ),
-                            WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/2DzJX8r.png",
-                            ),
-                          ],
-                        ),
-                        HeightSizedBox(targetSize: deviceHeight, value: 0.02),
-                        Container(
-                          color: const Color.fromRGBO(0, 0, 0, 0.3),
-                          height: 0.1,
-                          width: deviceWidth * 0.47,
-                        ),
-                        HeightSizedBox(targetSize: deviceHeight, value: 0.02),
-                        BodyText(
-                          text: "After",
-                          color: Colors.black,
-                          fontSize: deviceHeight * 0.03,
-                          fontWeight: FontWeight.bold,
-                          fontFamily: "Noto Snas JP",
-                        ),
-                        Row(
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          children: [
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/zbpIv7k.png",
-                            ),
-                            WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/ZZd5lXL.png",
-                            ),
-                            WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/zUphl8Y.png",
-                            ),
-                            WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                            const ImageWidget(
-                              heightValue: 0.33,
-                              imagePath: "https://i.imgur.com/0LAsoFt.png",
-                            ),
-                          ],
+                        HeightSizedBox(targetSize: deviceHeight, value: 0.01),
+                        const ImageWidget(
+                          heightValue: 0.7,
+                          imagePath: "https://i.imgur.com/mOshwq9.png",
                         ),
                       ],
                     ),
                     WidthSizedBox(targetSize: deviceWidth, value: 0.05),
                     Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      mainAxisAlignment: MainAxisAlignment.start,
+                      crossAxisAlignment: CrossAxisAlignment.center,
                       children: [
-                        HeightSizedBox(targetSize: deviceHeight, value: 0.03),
-                        Row(
-                          children: [
-                            const ImageWidget(
-                              heightValue: 0.45,
-                              imagePath: "https://i.imgur.com/zbpIv7k.png",
-                            ),
-                            WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                            const ImageWidget(
-                              heightValue: 0.45,
-                              imagePath: "https://i.imgur.com/zUphl8Y.png",
-                            ),
-                          ],
+                        BodyText(
+                          text: "オブジェクトモデリング",
+                          color: Colors.black,
+                          fontSize: deviceHeight * 0.03,
+                          fontWeight: FontWeight.bold,
+                          fontFamily: "Noto Snas JP",
                         ),
-                        HeightSizedBox(targetSize: deviceHeight, value: 0.03),
-                        HighPaddingText(
-                          text:
-                              "・一目見て分かる程ひどいタスクベースな画面から\n　オブジェクトとアクションを取り出し、まとめ上げたことにより\n　画面遷移数を大幅に減少させることが出来ました。",
-                          color: const Color.fromRGBO(0, 0, 0, 0.8),
-                          fontSize: deviceHeight * 0.02,
-                          fontWeight: FontWeight.normal,
-                          fontFamily: "Noto Sans JP",
-                          textAlign: TextAlign.start,
-                          paddingValue: 1.3,
+                        HeightSizedBox(targetSize: deviceHeight, value: 0.01),
+                        const ImageWidget(
+                          heightValue: 0.4,
+                          imagePath: "https://i.imgur.com/B8ooAYf.png",
                         ),
                         HeightSizedBox(targetSize: deviceHeight, value: 0.02),
-                        HighPaddingText(
-                          text:
-                              "・分散していた組織情報を一画面に集約し、入館のかなめとなる\n　QRコードを常に表示することで、入館スピードの向上を図りました。",
-                          color: const Color.fromRGBO(0, 0, 0, 0.8),
-                          fontSize: deviceHeight * 0.02,
-                          fontWeight: FontWeight.normal,
-                          fontFamily: "Noto Sans JP",
-                          textAlign: TextAlign.start,
-                          paddingValue: 1.3,
+                        BodyText(
+                          text: "ビューとナビゲーションの検討",
+                          color: Colors.black,
+                          fontSize: deviceHeight * 0.03,
+                          fontWeight: FontWeight.bold,
+                          fontFamily: "Noto Snas JP",
                         ),
-                        HeightSizedBox(targetSize: deviceHeight, value: 0.02),
-                        HighPaddingText(
-                          text:
-                              "・優先度はそれほど高くない、組織構成のそれぞれの要素を\n　絞り込み機能で切り替えることによって、余計なタブを増やさずに\n　組織オブジェクト内に納めることが出来ました。",
-                          color: const Color.fromRGBO(0, 0, 0, 0.8),
-                          fontSize: deviceHeight * 0.02,
-                          fontWeight: FontWeight.normal,
-                          fontFamily: "Noto Sans JP",
-                          textAlign: TextAlign.start,
-                          paddingValue: 1.3,
+                        HeightSizedBox(targetSize: deviceHeight, value: 0.01),
+                        const ImageWidget(
+                          heightValue: 0.25,
+                          imagePath: "https://i.imgur.com/NOlXk6Y.png",
                         ),
                       ],
                     ),
