+++ lib/view/web_display/works_pages_web/tomony_pages_web/tomony_six_web.dart
@@ -136,21 +136,14 @@ class TomonySix extends StatelessWidget {
                                       fontFamily: "Noto Sans JP",
                                     ),
                                     BodyText(
-                                      text: "大学の友達と遊びに出かけたり、飲み会によく行ったりする。",
+                                      text: "大学の友達と遊びに出かけたり、飲み会によく行く。",
                                       color: Colors.black,
                                       fontSize: deviceHeight * 0.02,
                                       fontWeight: FontWeight.normal,
                                       fontFamily: "Noto Sans JP",
                                     ),
                                     BodyText(
-                                      text: "大学1年の頃から付き合っている一個上で違う専攻の彼女が",
-                                      color: Colors.black,
-                                      fontSize: deviceHeight * 0.02,
-                                      fontWeight: FontWeight.normal,
-                                      fontFamily: "Noto Sans JP",
-                                    ),
-                                    BodyText(
-                                      text: "いる。普段は家で遊んだり、週末に出かけたりする。",
+                                      text: "大学1年の頃から付き合っている一個上の彼女がいる。",
                                       color: Colors.black,
                                       fontSize: deviceHeight * 0.02,
                                       fontWeight: FontWeight.normal,
@@ -205,21 +198,21 @@ class TomonySix extends StatelessWidget {
                                       fontFamily: "Noto Sans JP",
                                     ),
                                     BodyText(
-                                      text: "喧嘩することが増えてきた。彼女を好きな気持ちはあるが",
+                                      text: "喧嘩することが増えてきた。まだ彼女の分からない",
                                       color: Colors.black,
                                       fontSize: deviceHeight * 0.02,
                                       fontWeight: FontWeight.normal,
                                       fontFamily: "Noto Sans JP",
                                     ),
                                     BodyText(
-                                      text: "まだ彼女の分からないことも多く、彼女が求めている行動が",
+                                      text: "ことも多く、彼女が求めている行動ができず",
                                       color: Colors.black,
                                       fontSize: deviceHeight * 0.02,
                                       fontWeight: FontWeight.normal,
                                       fontFamily: "Noto Sans JP",
                                     ),
                                     BodyText(
-                                      text: "できず、どうすれば良い関係を築けるのか悩んでいる。",
+                                      text: "どうすれば良い関係を築けるのか悩んでいる。",
                                       color: Colors.black,
                                       fontSize: deviceHeight * 0.02,
                                       fontWeight: FontWeight.normal,
