+++ lib/view/web_display/works_pages_web/tomony_pages_web/tomony_three_web.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 
-import '../../../../utils/project_widget/about_widget.dart';
 import '../../../../../utils/extract_widget.dart';
 import '../../../../utils/project_widget/works_widget.dart';
 
@@ -22,11 +21,11 @@ class TomonyThree extends StatelessWidget {
         mainAxisAlignment: MainAxisAlignment.center,
         crossAxisAlignment: CrossAxisAlignment.center,
         children: <Widget>[
-          Column(
+          Row(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               SizedBox(
-                width: deviceWidth * 0.9,
+                height: deviceHeight * 0.59,
                 child: Container(
                   decoration: BoxDecoration(
                     borderRadius: BorderRadius.circular(10),
@@ -46,7 +45,7 @@ class TomonyThree extends StatelessWidget {
                         bottom: deviceHeight * 0.02,
                         left: deviceHeight * 0.03,
                         right: deviceHeight * 0.03),
-                    child: Row(
+                    child: Column(
                       mainAxisAlignment: MainAxisAlignment.center,
                       crossAxisAlignment: CrossAxisAlignment.center,
                       children: [
@@ -74,356 +73,250 @@ class TomonyThree extends StatelessWidget {
                                   ),
                                   HeightSizedBox(
                                       targetSize: deviceHeight, value: 0.01),
-                                  BodyText(
+                                  HighPaddingText(
                                     text:
-                                        "このサービスは、自分の彼女が生理中の時に、よく接し方や喧嘩の解決策を検索することから発案しました。",
+                                        "このサービスは、自分が彼女が生理中の時に、よく接し方や喧嘩の解決策を\n検索することから発案しました。問題を客観視できずに喧嘩に発展したり\n不満を溜め込んでしまうのではないかと仮定し、女性だけの問題ではなく\nそれ以上に男性の考え方・接し方を深めることが重要だと考え制作に至りました。",
                                     color: const Color.fromRGBO(0, 0, 0, 0.6),
                                     fontSize: deviceHeight * 0.02,
                                     fontWeight: FontWeight.normal,
                                     fontFamily: "Noto Sans JP",
-                                  ),
-                                  BodyText(
-                                    text:
-                                        "客観的に問題を見れずに喧嘩になってしまったり、溜め込んだりしてしまうのではないかと仮定し",
-                                    color: const Color.fromRGBO(0, 0, 0, 0.6),
-                                    fontSize: deviceHeight * 0.02,
-                                    fontWeight: FontWeight.normal,
-                                    fontFamily: "Noto Sans JP",
-                                  ),
-                                  BodyText(
-                                    text:
-                                        "女性だけの問題ではなく、それ以上に男性の考え方・接し方を深めることが重要だと考え制作に至りました。",
-                                    color: const Color.fromRGBO(0, 0, 0, 0.6),
-                                    fontSize: deviceHeight * 0.02,
-                                    fontWeight: FontWeight.normal,
-                                    fontFamily: "Noto Sans JP",
+                                    paddingValue: 1.3,
+                                    textAlign: TextAlign.left,
                                   ),
                                 ],
                               ),
                             ),
-                            BodyText(
-                              text: "前提",
-                              color: const Color(0xFF87C495),
-                              fontSize: deviceHeight * 0.028,
-                              fontWeight: FontWeight.bold,
-                              fontFamily: "源ノ角ゴシック VF",
-                            ),
-                            Padding(
-                              padding: EdgeInsets.all(deviceHeight * 0.02),
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: [
-                                  ShadowContainerText(
-                                    deviceHeight: deviceHeight,
-                                    text: "ピンポイントで自分達の状況にあった質問が少ない",
-                                  ),
-                                  HeightSizedBox(
-                                      targetSize: deviceHeight, value: 0.01),
-                                  ShadowContainerText(
-                                    deviceHeight: deviceHeight,
-                                    text: '専門的に1つにまとまっているサービスがない',
-                                  ),
-                                ],
-                              ),
-                            )
                           ],
                         ),
-                        WidthSizedBox(targetSize: deviceWidth, value: 0.03),
-                        const ImageWidget(
-                          heightValue: 0.38,
-                          imagePath: "https://i.imgur.com/PNGN244.png",
+                        Row(
+                          children: [
+                            Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                BodyText(
+                                  text: "前提",
+                                  color: const Color(0xFF87C495),
+                                  fontSize: deviceHeight * 0.028,
+                                  fontWeight: FontWeight.bold,
+                                  fontFamily: "源ノ角ゴシック VF",
+                                ),
+                                Padding(
+                                  padding: EdgeInsets.all(deviceHeight * 0.02),
+                                  child: Column(
+                                    crossAxisAlignment:
+                                        CrossAxisAlignment.start,
+                                    children: [
+                                      ShadowContainerText(
+                                        deviceHeight: deviceHeight,
+                                        text: "ピンポイントで自分達の状況にあった質問が少ない",
+                                      ),
+                                      HeightSizedBox(
+                                          targetSize: deviceHeight,
+                                          value: 0.015),
+                                      ShadowContainerText(
+                                        deviceHeight: deviceHeight,
+                                        text: '専門的に1つにまとまっているサービスがない',
+                                      ),
+                                    ],
+                                  ),
+                                )
+                              ],
+                            ),
+                            const ImageWidget(
+                              heightValue: 0.2,
+                              imagePath: "https://i.imgur.com/QL6xkNp.png",
+                            ),
+                          ],
                         ),
                       ],
                     ),
                   ),
                 ),
               ),
-              SizedBox(
-                width: deviceWidth * 0.92,
-                child: Padding(
-                  padding: EdgeInsets.only(
-                      top: deviceHeight * 0.03,
-                      right: deviceHeight * 0.03,
-                      left: deviceHeight * 0.03),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          BodyText(
-                            text: "リサーチ・インタビュー",
-                            color: const Color(0xFF87C495),
-                            fontSize: deviceHeight * 0.028,
-                            fontWeight: FontWeight.bold,
-                            fontFamily: "源ノ角ゴシック VF",
+              Padding(
+                padding: EdgeInsets.symmetric(
+                  horizontal: deviceHeight * 0.03,
+                ),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    Column(
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        SizedBox(
+                          width: deviceWidth * 0.35,
+                          child: Container(
+                            decoration: BoxDecoration(
+                              borderRadius: BorderRadius.circular(10),
+                              color: const Color.fromRGBO(151, 151, 151, 0.1),
+                            ),
+                            child: Padding(
+                              padding: EdgeInsets.all(deviceHeight * 0.035),
+                              child: Column(
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                children: [
+                                  BodyText(
+                                    text: "課題発見前の仮説",
+                                    color: Colors.black,
+                                    fontSize: deviceHeight * 0.028,
+                                    fontWeight: FontWeight.bold,
+                                    fontFamily: "Noto Sans JP",
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.025),
+                                  BodyText(
+                                    text: "悩んで相談したり、愚痴を言ったりしているのではないか？",
+                                    color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                    fontSize: deviceHeight * 0.02,
+                                    fontWeight: FontWeight.normal,
+                                    fontFamily: "Noto Sans JP",
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.02),
+                                  BodyText(
+                                    text: "相談や既存の回答を参考に、解決できた回数が少ないのでは？",
+                                    color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                    fontSize: deviceHeight * 0.02,
+                                    fontWeight: FontWeight.normal,
+                                    fontFamily: "Noto Sans JP",
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.02),
+                                  BodyText(
+                                    text: "あまり検索したり、彼女と生理の話をしていないのでは？",
+                                    color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                    fontSize: deviceHeight * 0.02,
+                                    fontWeight: FontWeight.normal,
+                                    fontFamily: "Noto Sans JP",
+                                  ),
+                                ],
+                              ),
+                            ),
                           ),
-                          Padding(
-                            padding: EdgeInsets.only(left: deviceHeight * 0.03),
-                            child: Row(
-                              children: [
-                                Column(
-                                  children: const [
-                                    TrueCircle(
-                                      sizeValue: 0.015,
-                                      color: Color(0xFF87C495),
-                                    ),
-                                    VerticalLine(
-                                      heightPadding: 0.01,
-                                      heightValue: 0.05,
-                                      lineColor: Color(0xFF87C495),
-                                    ),
-                                    TrueCircle(
-                                      sizeValue: 0.015,
-                                      color: Color(0xFF87C495),
-                                    ),
-                                    VerticalLine(
-                                      heightPadding: 0.01,
-                                      heightValue: 0.05,
-                                      lineColor: Color(0xFF87C495),
-                                    ),
-                                    TrueCircle(
-                                      sizeValue: 0.015,
-                                      color: Color(0xFF87C495),
-                                    ),
-                                    VerticalLine(
-                                      heightPadding: 0.01,
-                                      heightValue: 0.05,
-                                      lineColor: Color(0xFF87C495),
-                                    ),
-                                    TrueCircle(
-                                      sizeValue: 0.015,
-                                      color: Color(0xFF87C495),
-                                    ),
-                                    VerticalLine(
-                                      heightPadding: 0.01,
-                                      heightValue: 0.05,
-                                      lineColor: Color(0xFF87C495),
-                                    ),
-                                    TrueCircle(
-                                      sizeValue: 0.015,
-                                      color: Color(0xFF87C495),
-                                    ),
-                                  ],
-                                ),
-                                WidthSizedBox(
-                                    targetSize: deviceWidth, value: 0.01),
-                                SizedBox(
-                                  height: deviceHeight * 0.34,
-                                  child: Column(
-                                    mainAxisAlignment:
-                                        MainAxisAlignment.spaceEvenly,
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: const [
-                                      ProcessDetail(
-                                        process: "ターゲットと仮説立案",
-                                        detail: "課題前の仮説を元にインタビュー作成",
+                        ),
+                        HeightSizedBox(targetSize: deviceHeight, value: 0.02),
+                        SizedBox(
+                          width: deviceWidth * 0.35,
+                          child: Container(
+                            decoration: BoxDecoration(
+                              borderRadius: BorderRadius.circular(10),
+                              color: const Color.fromRGBO(151, 151, 151, 0.1),
+                            ),
+                            child: Padding(
+                              padding: EdgeInsets.all(deviceHeight * 0.03),
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.center,
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  BodyText(
+                                    text: "リサーチ結果(生理経験なし 400名)",
+                                    color: Colors.black,
+                                    fontSize: deviceHeight * 0.028,
+                                    fontWeight: FontWeight.bold,
+                                    fontFamily: "Noto Sans JP",
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.02),
+                                  Row(
+                                    mainAxisAlignment: MainAxisAlignment.center,
+                                    children: [
+                                      BodyText(
+                                        text: "生理についてよく理解している",
+                                        color:
+                                            const Color.fromRGBO(1, 1, 1, 0.8),
+                                        fontSize: deviceHeight * 0.02,
+                                        fontWeight: FontWeight.normal,
+                                        fontFamily: "Noto Sans JP",
                                       ),
-                                      ProcessDetail(
-                                        process: "リサーチ(インターネット)",
-                                        detail: "生理経験なしの若者の悩みを調査",
-                                      ),
-                                      ProcessDetail(
-                                        process: "インタビュー項目の検討",
-                                        detail: "本質を引き出すためのオープンクエスチョン",
-                                      ),
-                                      ProcessDetail(
-                                        process: "インタビュー",
-                                        detail: "学校の友人・スタッフ数名に実施",
-                                      ),
-                                      ProcessDetail(
-                                        process: "分析",
-                                        detail: "結果をもとに課題を抽出",
+                                      WidthSizedBox(
+                                          targetSize: deviceWidth, value: 0.01),
+                                      BodyText(
+                                        text: "21.1%",
+                                        color: const Color(0xFF87C495),
+                                        fontSize: deviceHeight * 0.025,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
                                       ),
                                     ],
                                   ),
-                                ),
-                              ],
-                            ),
-                          )
-                        ],
-                      ),
-                      WidthSizedBox(targetSize: deviceWidth, value: 0.03),
-                      SizedBox(
-                        height: deviceHeight * 0.35,
-                        child: Container(
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(10),
-                            color: const Color.fromRGBO(151, 151, 151, 0.1),
-                          ),
-                          child: Padding(
-                            padding: EdgeInsets.all(deviceHeight * 0.03),
-                            child: Column(
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              children: [
-                                BodyText(
-                                  text: "課題発見前の仮説",
-                                  color: Colors.black,
-                                  fontSize: deviceHeight * 0.028,
-                                  fontWeight: FontWeight.bold,
-                                  fontFamily: "Noto Sans JP",
-                                ),
-                                HeightSizedBox(
-                                    targetSize: deviceHeight, value: 0.025),
-                                BodyText(
-                                  text: "悩んで相談したり、愚痴を言ったりしているのではないか？",
-                                  color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                  fontSize: deviceHeight * 0.02,
-                                  fontWeight: FontWeight.normal,
-                                  fontFamily: "Noto Sans JP",
-                                ),
-                                HeightSizedBox(
-                                    targetSize: deviceHeight, value: 0.02),
-                                BodyText(
-                                  text: "相談や既存の回答を参考に、解決できた回数が少ないのでは？",
-                                  color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                  fontSize: deviceHeight * 0.02,
-                                  fontWeight: FontWeight.normal,
-                                  fontFamily: "Noto Sans JP",
-                                ),
-                                HeightSizedBox(
-                                    targetSize: deviceHeight, value: 0.02),
-                                BodyText(
-                                  text: "あまり検索したり、彼女と生理の話をしていないのでは？",
-                                  color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                  fontSize: deviceHeight * 0.02,
-                                  fontWeight: FontWeight.normal,
-                                  fontFamily: "Noto Sans JP",
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                      ),
-                      WidthSizedBox(targetSize: deviceWidth, value: 0.01),
-                      SizedBox(
-                        height: deviceHeight * 0.35,
-                        child: Container(
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(10),
-                            color: const Color.fromRGBO(151, 151, 151, 0.1),
-                          ),
-                          child: Padding(
-                            padding: EdgeInsets.all(deviceHeight * 0.03),
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.center,
-                              mainAxisAlignment: MainAxisAlignment.center,
-                              children: [
-                                BodyText(
-                                  text: "リサーチ結果(生理経験なし 400名)",
-                                  color: Colors.black,
-                                  fontSize: deviceHeight * 0.028,
-                                  fontWeight: FontWeight.bold,
-                                  fontFamily: "Noto Sans JP",
-                                ),
-                                HeightSizedBox(
-                                    targetSize: deviceHeight, value: 0.02),
-                                Column(
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  children: [
-                                    Row(
-                                      children: [
-                                        BodyText(
-                                          text: "生理についてよく理解している",
-                                          color: const Color.fromRGBO(
-                                              1, 1, 1, 0.8),
-                                          fontSize: deviceHeight * 0.02,
-                                          fontWeight: FontWeight.normal,
-                                          fontFamily: "Noto Sans JP",
-                                        ),
-                                        WidthSizedBox(
-                                            targetSize: deviceWidth,
-                                            value: 0.01),
-                                        BodyText(
-                                          text: "21.1%",
-                                          color: const Color(0xFF87C495),
-                                          fontSize: deviceHeight * 0.025,
-                                          fontWeight: FontWeight.bold,
-                                          fontFamily: "Noto Sans JP",
-                                        ),
-                                      ],
-                                    ),
-                                    HeightSizedBox(
-                                        targetSize: deviceHeight, value: 0.01),
-                                    Row(
-                                      children: [
-                                        BodyText(
-                                          text: "生理の話をできる相談相手がいる",
-                                          color: const Color.fromRGBO(
-                                              1, 1, 1, 0.8),
-                                          fontSize: deviceHeight * 0.02,
-                                          fontWeight: FontWeight.normal,
-                                          fontFamily: "Noto Sans JP",
-                                        ),
-                                        WidthSizedBox(
-                                            targetSize: deviceWidth,
-                                            value: 0.01),
-                                        BodyText(
-                                          text: "34.2%",
-                                          color: const Color(0xFF87C495),
-                                          fontSize: deviceHeight * 0.025,
-                                          fontWeight: FontWeight.bold,
-                                          fontFamily: "Noto Sans JP",
-                                        ),
-                                      ],
-                                    ),
-                                    HeightSizedBox(
-                                        targetSize: deviceHeight, value: 0.01),
-                                    Row(
-                                      children: [
-                                        LongText(
-                                          text:
-                                              "理解不足を感じ、生理について\nもっと理解したいと感じている",
-                                          color: const Color.fromRGBO(
-                                              1, 1, 1, 0.8),
-                                          fontSize: deviceHeight * 0.02,
-                                          fontWeight: FontWeight.normal,
-                                          fontFamily: "Noto Sans JP",
-                                          textAlign: TextAlign.start,
-                                        ),
-                                        WidthSizedBox(
-                                            targetSize: deviceWidth,
-                                            value: 0.01),
-                                        BodyText(
-                                          text: "約半数",
-                                          color: const Color(0xFF87C495),
-                                          fontSize: deviceHeight * 0.025,
-                                          fontWeight: FontWeight.bold,
-                                          fontFamily: "Noto Sans JP",
-                                        ),
-                                      ],
-                                    ),
-                                  ],
-                                ),
-                                HeightSizedBox(
-                                    targetSize: deviceHeight, value: 0.01),
-                                BodyText(
-                                  text: "⬇︎",
-                                  color: const Color.fromRGBO(1, 1, 1, 0.8),
-                                  fontSize: deviceHeight * 0.02,
-                                  fontWeight: FontWeight.bold,
-                                  fontFamily: "Noto Sans JP",
-                                ),
-                                HeightSizedBox(
-                                    targetSize: deviceHeight, value: 0.01),
-                                HighPaddingText(
-                                  text: "・生理の人に対する接し方を知りたい\n・生理への向き合い方を知りたい",
-                                  color: const Color(0xFF87C495),
-                                  fontSize: deviceHeight * 0.02,
-                                  fontWeight: FontWeight.bold,
-                                  fontFamily: "Noto Sans JP",
-                                  textAlign: TextAlign.center,
-                                  paddingValue: 1.2,
-                                ),
-                              ],
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.01),
+                                  Row(
+                                    mainAxisAlignment: MainAxisAlignment.center,
+                                    children: [
+                                      BodyText(
+                                        text: "生理の話をできる相談相手がいる",
+                                        color:
+                                            const Color.fromRGBO(1, 1, 1, 0.8),
+                                        fontSize: deviceHeight * 0.02,
+                                        fontWeight: FontWeight.normal,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                      WidthSizedBox(
+                                          targetSize: deviceWidth, value: 0.01),
+                                      BodyText(
+                                        text: "34.2%",
+                                        color: const Color(0xFF87C495),
+                                        fontSize: deviceHeight * 0.025,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                    ],
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.01),
+                                  Row(
+                                    mainAxisAlignment: MainAxisAlignment.center,
+                                    children: [
+                                      LongText(
+                                        text: "理解不足を感じ、生理について\nもっと理解したいと感じている",
+                                        color:
+                                            const Color.fromRGBO(1, 1, 1, 0.8),
+                                        fontSize: deviceHeight * 0.02,
+                                        fontWeight: FontWeight.normal,
+                                        fontFamily: "Noto Sans JP",
+                                        textAlign: TextAlign.start,
+                                      ),
+                                      WidthSizedBox(
+                                          targetSize: deviceWidth, value: 0.01),
+                                      BodyText(
+                                        text: "約半数",
+                                        color: const Color(0xFF87C495),
+                                        fontSize: deviceHeight * 0.025,
+                                        fontWeight: FontWeight.bold,
+                                        fontFamily: "Noto Sans JP",
+                                      ),
+                                    ],
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.01),
+                                  BodyText(
+                                    text: "⬇︎",
+                                    color: const Color.fromRGBO(1, 1, 1, 0.8),
+                                    fontSize: deviceHeight * 0.02,
+                                    fontWeight: FontWeight.bold,
+                                    fontFamily: "Noto Sans JP",
+                                  ),
+                                  HeightSizedBox(
+                                      targetSize: deviceHeight, value: 0.01),
+                                  HighPaddingText(
+                                    text: "・生理の人に対する接し方を知りたい\n・生理への向き合い方を知りたい",
+                                    color: const Color(0xFF87C495),
+                                    fontSize: deviceHeight * 0.02,
+                                    fontWeight: FontWeight.bold,
+                                    fontFamily: "Noto Sans JP",
+                                    textAlign: TextAlign.center,
+                                    paddingValue: 1.2,
+                                  ),
+                                ],
+                              ),
                             ),
                           ),
-                        ),
-                      ),
-                    ],
-                  ),
+                        )
+                      ],
+                    ),
+                  ],
                 ),
               ),
             ],
